pyproject.toml
[project]
name = "alec"
version = "0.1.0"
description = "ALEC Language Learning Crew using CrewAI with Gemini Flash"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.14"
dependencies = [
    "crewai[tools]>=0.141.0,<1.0.0",
    "weave>=0.50.0",
    "wandb>=0.16.0",
    "python-dotenv>=1.0.0"
]

[project.scripts]
alec = "alec.main:run"
run_crew = "alec.main:run"
train = "alec.main:train"
replay = "alec.main:replay"
test = "alec.main:test"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"